fix: separate web, native
